wevtutil.exe deployment for log evasion
Wevtutil.exe is created by windows to be used by system administrator's and for mainly trouble shooting purposes; not for actors to abuse to hide their tracks or create visibility gaps. The binary really shines less when used to straight up clear logs, as it is useful to momentarily disable logs of interest to preform certain actions more stealthy and then resume logging as normal.
It is unlikely that defenders would notice a 2-3 minute gap in logging and thus can allow you crucial minutes to preform nosier/high risk activity while reducing risk of detection. Many red teamers and pentesters(at my workplace included) were not aware of the ability to pause logging and resume it with relative ease and expressed interest in having it handy. [Potentially others are in the same spot and would benefit from having it in the project!